order process finished

+++ ecommerce/store/models.py
@@ -14,9 +14,9 @@ class Customer(models.Model):
 # Clasa Product, care reprezintă modelele produselor disponibile în magazin.
 class Product(models.Model):
     name    = models.CharField(max_length=200)
-    price   = models.FloatField()
+    price   = models.DecimalField(max_digits=7,decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=True)
-    image=models.ImageField(null=True, blank=True)
+    image   =models.ImageField(null=True, blank=True)
 
     def __str__(self) :
         return self.name
